Fixed routes so that the main file defaults to the login page
